requirements.txt
autopep8==1.5.7
click==8.0.1
dnspython==2.1.0
Flask==2.0.1
Flask-PyMongo==2.3.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pycodestyle==2.7.0
pymongo==3.11.4
python-dotenv==0.17.1
toml==0.10.2
Werkzeug==2.0.1
